debug

+++ src/main/java/com/codecool/order/Order.java
@@ -11,7 +11,7 @@ public class Order implements Orderable {
     private String status;
     private boolean isPaid;
     private boolean isChecked;
-    private int orderValue;
+    private float orderValue;
 
     public Order() {
         this.id = 100 + orderCounter;
@@ -30,7 +30,7 @@ public class Order implements Orderable {
         return this.status;
     }
 
-    public int getOrderValue() {
+    public float getOrderValue() {
         return this.orderValue;
     }
 
@@ -38,7 +38,7 @@ public class Order implements Orderable {
         return basket;
     }
 
-    public void setOrderValue(int newValue) {
+    public void setOrderValue(float newValue) {
         this.orderValue = newValue;
     }
